+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Calling **format** with the `-d` flag now removes test playbooks testing the deprecated content from conf.json.
 * Fixed an issue where in some cases the **split** command did not remove pack version note from the script.
 * Fixed an issue where **validate** would not properly detect dependencies of core packs.
+* Removed usages of Random in unit tests to ensure the tests are deterministic.
 * **validate** will now run on all the pack content items when the pack supported marketplaces are modified.
 
 ## 1.17.2